+++ src/_utils/branding.unit.ts
@@ -1,0 +1,5 @@
+import { pxToInteger } from './branding'
+
+it('should convert a string px value to integer', () => {
+  expect(pxToInteger('10px')).toEqual(10)
+})