+++ src/stepper/story.tsx
@@ -2,9 +2,10 @@ import React from 'react'
 
 import { storiesOf } from '@storybook/react'
 import { action } from '@storybook/addon-actions'
-import { withKnobs, text, number } from '@storybook/addon-knobs'
+import { withKnobs, text, number, select } from '@storybook/addon-knobs'
 
-import Stepper from 'stepper'
+import spec from './specifications/stepper.md'
+import Stepper, { StepperDisplay } from 'stepper'
 
 const stories = storiesOf('Stepper', module)
 stories.addDecorator(withKnobs)
@@ -19,11 +20,13 @@ stories.add('Default stepper', () => (
     increaseLabel={text('increaseLabel', 'Increment')}
     decreaseLabel={text('decreaseLabel', 'Decrement')}
     onChange={action('changed')}
-    buttonSize={number('button size', 24)}
+    display={select('display', StepperDisplay, StepperDisplay.SMALL)}
   >
     {text('children', 'Number of seats')}
   </Stepper>
-))
+), {
+  readme: { content: spec },
+},)
 
 stories.add('Stepper with formatted value', () => (
   <Stepper
@@ -36,7 +39,24 @@ stories.add('Stepper with formatted value', () => (
     decreaseLabel={text('decreaseLabel', 'Decrement')}
     format={value => `${value} €`}
     onChange={action('changed')}
-    buttonSize={number('button size', 24)}
+    display={select('display', StepperDisplay, StepperDisplay.SMALL)}
+  >
+    {text('children', 'Edit the price')}
+  </Stepper>
+))
+
+stories.add('Stepper with large formatted value', () => (
+  <Stepper
+    name="stepper2"
+    min={number('min', 0)}
+    max={number('max', 99999)}
+    step={number('step', 100)}
+    value={number('value', 8400)}
+    increaseLabel={text('increaseLabel', 'Increment')}
+    decreaseLabel={text('decreaseLabel', 'Decrement')}
+    format={value => `${value} грн.`}
+    onChange={action('changed')}
+    display={select('display', StepperDisplay, StepperDisplay.LARGE)}
   >
     {text('children', 'Edit the price')}
   </Stepper>