+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Unreleased
 
+- **[BREAKING CHANGE]** `display` replaces `buttonSize` prop in `Stepper` component
+- **[NEW]** `Stepper` component now renders its value depending on available space
 - **[NEW]** Focus the select when `focus` prop changes on `TimePicker`
 - **[FIX]** Fix right margin on `MarketingMessage`
+[...]
 
 # v14.0.1 (08/11/2019)